requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.12.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
docopt==0.6.2
Flask==2.2.3
frozenlist==1.3.3
geographiclib==2.0
geopy==2.3.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Js2Py==0.74
MarkupSafe==2.1.2
MouseInfo==0.1.3
multidict==6.0.4
openai==0.27.2
packaging==23.0
Pillow==9.4.0
PyAudio==0.2.13
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
pyjokes==0.6.0
pyjsparser==2.7.1
PyMsgBox==1.0.9
pyperclip==1.8.2
PyPrind==2.11.3
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.1
PyQt6==6.4.2
PyQt6-Qt6==6.4.3
PyQt6-sip==13.4.1
PyQt6-WebEngine==6.4.0
PyQt6-WebEngine-Qt6==6.4.3
PyRect==0.2.0
PyScreeze==0.1.28
pySmartDL==1.3.4
python-dotenv==1.0.0
python3-xlib==0.15
pyttsx3==2.90
pytweening==1.0.4
pytz-deprecation-shim==0.1.0.post0
pywhatkit==5.4
requests==2.28.2
six==1.16.0
soupsieve==2.4
SpeechRecognition==3.10.0
tqdm==4.65.0
tzdata==2022.7
tzlocal==4.3
urllib3==1.26.15
Werkzeug==2.2.3
wikipedia==1.4.0
yarl==1.8.2
